// src/formsClient/models/FieldConditionRuleOperator.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum FieldConditionRuleOperator {
    IS = 'Is',
    IS_NOT = 'IsNot',
    GREATER_THEN = 'GreaterThen',
    LESS_THEN = 'LessThen',
    CONTAINS = 'Contains',
    CONTAINS_IGNORE_CASE = 'ContainsIgnoreCase',
    STARTS_WITH = 'StartsWith',
    STARTS_WITH_IGNORE_CASE = 'StartsWithIgnoreCase',
    ENDS_WITH = 'EndsWith',
    ENDS_WITH_IGNORE_CASE = 'EndsWithIgnoreCase',
    NOT_CONTAINS = 'NotContains',
    NOT_CONTAINS_IGNORE_CASE = 'NotContainsIgnoreCase',
    NOT_STARTS_WITH = 'NotStartsWith',
    NOT_STARTS_WITH_IGNORE_CASE = 'NotStartsWithIgnoreCase',
    NOT_ENDS_WITH = 'NotEndsWith',
    NOT_ENDS_WITH_IGNORE_CASE = 'NotEndsWithIgnoreCase',
}