// src/formsClient/models/FormConditionRuleDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FieldConditionRuleOperator } from './FieldConditionRuleOperator';
export type FormConditionRuleDto = {
    field: string;
    operator: FieldConditionRuleOperator;
    value: string;
};